refactor: project restructure based on postgres schema hierarchy
