+++ src/support_sphere_db/src/ssdb/repositories/auth/__init__.py
@@ -1,0 +1,3 @@
+from ssdb.repositories.auth.user_repository import UserRepository
+
+__all__ = ['UserRepository']
