+++ src/support_sphere_db/src/ssdb/repositories/public/__init__.py
@@ -1,0 +1,3 @@
+from ssdb.repositories.public.user_profile_repository import UserProfileRepository
+
+__all__ = ['UserProfileRepository']
