--- src/support_sphere_db/src/ssdb/repository/public/__init__.py
@@ -1,3 +1,0 @@
-from ssdb.repository.public.user_profile_repository import UserProfileRepository
-
-__all__ = ['UserProfileRepository']
